Refactor: public api 적용 - #17
