+++ src/components/common/index.ts
@@ -1,0 +1,3 @@
+export * from './buttons';
+export * from './input';
+export * from './modal';
