+++ src/components/common/input/index.ts
@@ -1,0 +1,1 @@
+export * from './CommonInput';
