+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './auth';
+export * from './common';
+export * from './links';
+export * from './modals';
