+++ src/components/links/FolderList/Folder.tsx
@@ -2,12 +2,10 @@ interface FolderProps {
   folderName: string;
 }
 
-const Folder = ({ folderName }: FolderProps) => {
+export const Folder = ({ folderName }: FolderProps) => {
   return (
     <button className="px-3 py-2 border border-solid border-primary rounded-[5px] font-normal text-base font-[Pretendard] not-italic leading-[normal] hover:bg-[#e7effb]">
       {folderName}
     </button>
   );
 };
-
-export default Folder;
