+++ src/components/links/FolderList/index.ts
@@ -1,0 +1,2 @@
+export * from './Folder';
+export * from './FolderList';
