+++ src/components/links/FolderMenuList/index.ts
@@ -1,0 +1,2 @@
+export * from './FolderMenu';
+export * from './FolderMenuList';
